refactor: moving background into app
